#36 updated documentation and renamed some files
