+++ public/views/login.ejs
@@ -14,17 +14,16 @@
     <header>
         <%- include('./partials/header'); %>
     </header>
-             <div id="login-alert-container" class="fixed inset-x-0 top-[75px] z-50"></div>
+    <div id="login-alert-container" class="fixed inset-x-0 top-[75px] z-50"></div>
     <%- include('./partials/drawer/drawerLogin'); %>
         <div class="flex flex-col items-center justify-center px-6 py-8 mx-auto md:h-screen lg:py-0">
             <a href="#" class="flex items-center mb-6 text-2xl font-semibold text-gray-900 dark:text-white">
                 <img class="w-8 h-8 mr-2" src="../images/favicon_io/favicon.ico" alt="logo">
                 FORSETE
             </a>
-             <h1
-                        class="text-xl pb-4 font-bold leading-tight tracking-tight text-gray-900 md:text-2xl dark:text-white">
-                        Welcome! Convert handwritten or typed documents into editable text.
-                    </h1>
+            <h1 class="text-xl pb-4 font-bold leading-tight tracking-tight text-gray-900 md:text-2xl dark:text-white">
+                Welcome! Convert handwritten or typed documents into editable text.
+            </h1>
             <div
                 class="w-full bg-white rounded-lg shadow dark:border md:mt-0 sm:max-w-md xl:p-0 dark:bg-gray-800 dark:border-gray-700">
                 <div class="p-6 space-y-4 md:space-y-6 sm:p-8">
