+++ src/controllers/components/zoomImage.ts
